Added a Firestorm build.
Updating with a Firestorm build as well. I did some sims on my own side figuring out this build. If you wanna check with Preheat again, feel free to do so. I tried to contact them on discord, but I know how it is to have loads of people trying to contact you. So I wasn't expecting to be added. I tried to help Bloodmallet out by not having to take that extra step to contact him first.